+++ packages/plugin-packages/mashroom-http-proxy/src/connection_pool.ts
@@ -2,9 +2,10 @@
 import http from 'http';
 import https from 'https';
 
-import type {Agent as HttpAgent} from 'http';
+import type {Agent as HttpAgent,ClientRequest} from 'http';
 import type {Agent as HttpsAgent} from 'https';
 
+import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
 import type {PoolConfig, PoolMetrics} from '../type-definitions/internal';
 
 let _config: PoolConfig = {
@@ -44,26 +45,62 @@ export const getHttpsPool = () => {
     return _httpsPool;
 };
 
-const getPoolStats = (agent: HttpAgent | HttpsAgent): PoolMetrics => {
+const addTargetCount = (clientRequest: ClientRequest, hostCount: Record<string, number>, logger: MashroomLogger) => {
+    try {
+        const target = `${clientRequest.protocol}//${clientRequest.getHeader('host')}`;
+        if (hostCount[target]) {
+            hostCount[target] ++;
+        } else {
+            hostCount[target] = 1;
+        }
+    } catch (e) {
+        logger.error('Determining URL for ClientRequest failed', e);
+    }
+};
+
+const getPoolStats = (agent: HttpAgent | HttpsAgent, logger: MashroomLogger): PoolMetrics => {
     const countArrayEntries = (obj: NodeJS.ReadOnlyDict<any>) => Object.values(obj).reduce((acc, arr) => acc + arr.length, 0);
+
+    const activeConnectionsTargetCount: Record<string, number> = {};
+    Object.values(agent.sockets).forEach((activeSockets) => {
+        if (activeSockets) {
+            activeSockets.forEach((activeSocket) => {
+                // @ts-ignore
+                const clientRequest = activeSocket._httpMessage as ClientRequest;
+                addTargetCount(clientRequest, activeConnectionsTargetCount, logger);
+            });
+        }
+    });
+
+    const waitingRequestsTargetCount: Record<string, number> = {};
+    Object.values(agent.requests).forEach((waitingRequests) => {
+        if (waitingRequests) {
+            waitingRequests.forEach((waitingRequest) => {
+                const clientRequest = waitingRequest as unknown as ClientRequest;
+                addTargetCount(clientRequest, waitingRequestsTargetCount, logger);
+            });
+        }
+    });
 
     return {
         activeConnections: countArrayEntries(agent.sockets),
+        activeConnectionsTargetCount,
         idleConnections: countArrayEntries(agent.freeSockets),
         waitingRequests: countArrayEntries(agent.requests),
+        waitingRequestsTargetCount,
     };
 };
 
-export const getHttpPoolMetrics = (): PoolMetrics | null => {
+export const getHttpPoolMetrics = (logger: MashroomLogger): PoolMetrics | null => {
     if (_httpPool) {
-        return getPoolStats(_httpPool);
+        return getPoolStats(_httpPool, logger);
     }
     return null;
 };
 
-export const getHttpsPoolMetrics = (): PoolMetrics | null => {
+export const getHttpsPoolMetrics = (logger: MashroomLogger): PoolMetrics | null => {
     if (_httpsPool) {
-        return getPoolStats(_httpsPool);
+        return getPoolStats(_httpsPool, logger);
     }
     return null;
 };